# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tire Management System API",
        "description": "API documentation for the Tire Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/operations/vehicles": {
            "get": {
                "tags": [
                    "Operations"
                ],
                "summary": "List all vehicles",
                "operationId": "1c57cfddb11663a0f01746aa1f760949",
                "responses": {
                    "200": {
                        "description": "List of vehicles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Register a new vehicle",
                "operationId": "458c3bfea49b110ac53548c6ed216697",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "registration_number",
                                    "model",
                                    "type"
                                ],
                                "properties": {
                                    "registration_number": {
                                        "type": "string"
                                    },
                                    "fleet_number": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "axle_config": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created"
                    }
                }
            }
        },
        "/api/operations/vehicles/{id}": {
            "get": {
                "tags": [
                    "Operations"
                ],
                "summary": "Get vehicle details",
                "operationId": "2b9de26cb267360c4f370c4388665f87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle details with mounted tires"
                    }
                }
            },
            "put": {
                "tags": [
                    "Operations"
                ],
                "summary": "Update vehicle details",
                "operationId": "29273df9449449aaa984aec6b2f53371",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "model": {
                                        "type": "string"
                                    },
                                    "fleet_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated"
                    }
                }
            }
        },
        "/api/operations/vehicles/{id}/timeline": {
            "get": {
                "tags": [
                    "Operations"
                ],
                "summary": "Get vehicle operation timeline",
                "operationId": "25595d6ff8d6a7f3ed9b7641cd44766e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle operations timeline"
                    }
                }
            }
        },
        "/api/operations/vehicles/{id}/archive": {
            "delete": {
                "tags": [
                    "Operations"
                ],
                "summary": "Archive a vehicle",
                "operationId": "c74b79375fb30ca86f3e22b951888f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle archived"
                    }
                }
            }
        },
        "/api/operations/vehicles/{id}/retire": {
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Retire a vehicle",
                "operationId": "b42f97bf73fd02aba5c2f4d7bf51561b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle retired"
                    }
                }
            }
        },
        "/api/vehicles/{id}/inspections": {
            "get": {
                "tags": [
                    "Operations"
                ],
                "summary": "Get all inspections for a vehicle",
                "operationId": "d284f6b524bbc34e49b9793cb0d22768",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of inspections"
                    }
                }
            }
        },
        "/api/reports/low-stock": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Low stock report",
                "operationId": "910d470f508b24d03c3329ab9fb9f714",
                "responses": {
                    "200": {
                        "description": "Tires low in stock"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/tire-summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Tire status summary",
                "operationId": "97c9ffef3bf778537ed58696d6f0b893",
                "responses": {
                    "200": {
                        "description": "Summary of tires by status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/inspection-compliance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Inspection compliance report",
                "operationId": "fa4c0ce97f6bdc9128661ba314d10e60",
                "responses": {
                    "200": {
                        "description": "Compliance data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reports/tire-performance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Tire Performance Analysis (CPK)",
                "operationId": "b2ea157cb95e035bb153c09d12676dcb",
                "responses": {
                    "200": {
                        "description": "Performance metrics per brand/model"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}